server/requirements.txt
fastapi==0.103.1
uvicorn==0.29.0
python-dotenv==1.0.1
pymupdf==1.24.4
pypdf==4.2.0
langchain_community==0.2.0
unstructured==0.14.0
langchain==0.2.0
langchain-core==0.2.0
pdfminer==20191125
langchain-openai==0.1.7
tiktoken==0.7.0
chromadb==0.5.0